Prep and search issue
prepAndSearch function was being called for each block that was filled when it should only run once at the end.  I created a quick and dirty way to check when all blocks are complete and filled. Probably much more elegant ways of doing this with closures or something instead of globals.